+++ price-adapter-raw/CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Support CoinGecko sources
 - Support Binance Websocket source
+- Support Huobipro sources