+++ price-adapter/CHANGELOG.md
@@ -4,4 +4,5 @@
 
 - Support CoinGecko sources
 - Support Binance Websocket source
+- Support HuobiPro sources
 - Create Interval and Websocket service to manage sources